+++ RELEASE-NOTES.md
@@ -3,12 +3,14 @@
 ## 4.0
 **Breaking changes:**
 * When you specify null arguments like `request.WithArgument("x", null)`, the argument is now ignored instead of sending a blank value. You can use the previous behaviour by setting the `IgnoreNullArguments: false` option.
+* Removed deprecated methods.
 -->
 
 ## Upcoming release
 See [log](https://github.com/Pathoschild/FluentHttpClient/compare/3.2...develop).
 
 * Added support for reading responses into `JToken`, `JObject`, `JArray`, or `dynamic`.
+* Added fluent HTTP body builders, and deprecated some older body methods to simplify usage.
 * Added `IsSuccessStatusCode` to response.
 * Fixed default web proxy not being used.
 * Updated dependencies.